added option to output CIE perturbed initial conditions.
One caveat is that the CIE initial conditions currently assume
that we are evolving a single temperature because we have to define
which species obeys the conservation equation.  Perhaps this isn't
an issue, but it something to be aware of.  Also, I currently disabled
cooling for testing purposes due to the fact that oxygen won't evolve
for the entire time when it hits the temperature floor.  Still need
to determine a way of handling this.