modified reaction_classes to extrapolation in log-space.
for temperature ranges that exceed the ion-by-ion cooling tables
from Gnat, we now do a log-space extrapolation of the values.
This is not the best solution, but should avoid negative cooling
rates or sharp corners in the cooling function.  The long term
solution will likely require gather additional cooling rate
data.  I also thought this might help the solver keep going once
the gas cools to the oxygen temperature floor, but it does not.
A solution for handling that is still needed.