Converted oxygen species names to the roman numeral convention.
This was a little tricky and led to the need for some careful renaming
when calling ChiantiPy.  It also required using ChiantiPy in the
ion_cooling_rate function, which was not previously required.

I tried to do this in the best manner possible, but there may be a
cleaner/more efficient way to handle things.  At some point I'll
convert the code for Carbon (which should be relatively simple)

Also, this required modifying the species class to carry around
atomic number in addition to atomic weight.  I think that, in general,
this is a useful change regardless.  I also updated known_species.py
to acccount for this.