This adds a check in interpolate rates if the rate is a photo-heating rate.
Since the photoheat rates are redshift dependent, the interpolation
needs to be handled differently.  This also does a check and says that if the
redshift is outside the redshift for out photoheating table, then just
set the value of the photo heating to zero.

The same tricks should work for the photo-ionization rates.